+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.1] / 2023-11-10
 - Improve ordering of Format target dependencies
 
 ## [0.3.0] / 2023-11-09
@@ -30,7 +32,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.1.0] / 2023-11-08
 - Initial release
 
-[Unreleased]: https://github.com/vipentti/Vipentti.Nuke.Components/compare/0.3.0...HEAD
+[Unreleased]: https://github.com/vipentti/Vipentti.Nuke.Components/compare/0.3.1...HEAD
+[0.3.1]: https://github.com/vipentti/Vipentti.Nuke.Components/compare/0.3.0...0.3.1
 [0.3.0]: https://github.com/vipentti/Vipentti.Nuke.Components/compare/0.2.3...0.3.0
 [0.2.3]: https://github.com/vipentti/Vipentti.Nuke.Components/compare/0.2.2...0.2.3
 [0.2.2]: https://github.com/vipentti/Vipentti.Nuke.Components/compare/0.2.1...0.2.2